refactor(controller): move existing API to v1 package
fix #1
