+++ backend/src/com/atguigu/schedule/controller/v1/SysUserController.java
@@ -1,4 +1,4 @@
-package com.atguigu.schedule.controller;
+package com.atguigu.schedule.controller.v1;
 
 import com.atguigu.schedule.common.ResultCodeEnum;
 import com.atguigu.schedule.pojo.SysUser;
@@ -9,7 +9,6 @@ import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.JwtException;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
-import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
